fix: 🐛 Internal Representation of F32/F64

+++ src/structure/instructions.rs
@@ -8,8 +8,8 @@ pub enum Instr {
     /* Numeric Instructions */
     I32Const(i32),
     I64Const(i64),
-    F32Const(u32), // Bits of IEEE Binary 32 Floating Point Value
-    F64Const(u64), // Bits of IEEE Binary 64 Floating Point Value
+    F32Const(f32), // Bits of IEEE Binary 32 Floating Point Value
+    F64Const(f64), // Bits of IEEE Binary 64 Floating Point Value
     I32Clz,
     I32Ctz,
     I32Popcnt,
